debug: rename pallet for benchmark
