+++ runtime/litentry/src/weights/mod.rs
@@ -25,6 +25,7 @@ pub mod pallet_drop3;
 pub mod pallet_extrinsic_filter;
 pub mod pallet_membership;
 pub mod pallet_multisig;
+pub mod pallet_parachain_staking;
 pub mod pallet_preimage;
 pub mod pallet_proxy;
 pub mod pallet_scheduler;
@@ -32,4 +33,3 @@ pub mod pallet_session;
 pub mod pallet_timestamp;
 pub mod pallet_treasury;
 pub mod pallet_utility;
-pub mod parachain_staking;
